Added nuget packages and config for DAL support in Docker
Added nuget packages and config for DAL support in Docker by copying the nuget package manually over and then writing a nuget config file to use the relative path.  Now when the app builds in a docker container it has the nuget packages "locally"
Need to update the DAL build to copy new nuget packages over.